% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdrisk.R
\name{fdrisk}
\alias{fdrisk}
\title{False Discovery Risk for Second-Generation p-values}
\usage{
fdrisk(sgpvalue, null.lo, null.hi, pt.est, std.err,
  interval.type = "likelihood", interval.level = 1/8, pi0 = 0.5,
  null.weights, null.space, alt.weights, alt.space)
}
\arguments{
\item{sgpvalue}{The observed second-generation p-value}

\item{null.lo}{The lower bound for the indifference zone (null hypothesis) upon which the second-generation p-value was calculated.}

\item{null.hi}{The upper bound for the indifference zone (null hypothesis) upon which the second-generation p-value was calculated.}

\item{pt.est}{Point estimate for the parameter}

\item{std.err}{Standard error estimate for the point estimate}

\item{interval.type}{Type of interval estimate used. This determines the functional form of the power function. Options are \code{confidence} and \code{likelihood} (\code{credible} not yet supported).}

\item{interval.level}{Level of interval used. If \code{interval.type} is \code{confidence}, the alpha level, and if \code{interval.type} is \code{likelihood}, the quantity 1/k for a 1/k likelihood support interval.}

\item{pi0}{Prior probability that the null is true.}

\item{null.weights}{Type of probability distribution specified for the null parameter space. Options are currently \code{point} and \code{uniform}.}

\item{null.space}{Support for the null probability distribution. If \code{null.weights} is \code{point}, then \code{null.space} is a scalar, and if \code{null.weights} is \code{uniform}, then \code{null.space} is a vector of length two.}

\item{alt.weights}{Type of probability distribution specified for the alternative parameter space. Options are currently \code{point} and \code{uniform}.}

\item{alt.space}{Support for the alternative probability distribution. If \code{alt.weights} is \code{point}, then \code{alt.space} is a scalar, and if \code{alt.weights} is \code{uniform}, then \code{alt.space} is a vector of length two.}
}
\value{
Numeric. False discovery risk or false confirmation risk for the observed effect. If \code{sgpvalue} = 0, returns false discovery risk, and if \code{sgpvalue} = 1, returns false confirmation risk.
}
\description{
This function calculates the false discovery risk for a second-generation p-value finding of 0, or the false confirmation risk for a second-generation p-value finding of 1.
}
\details{
[Details]
Note that the second-generation \emph{p}-value and thus FDR or FCR should be calculated on the "symmetric" scale. For example, if the parameter of interest is an odds ratio, the inputs \code{pt.est}, \code{std.err}, \code{null.lo},  \code{null.hi}, \code{null.space}, and \code{alt.space} should all be on the scale of log odds ratios.
}
\examples{
# false discovery risk with confidence level
> fdrisk(sgpvalue = 0,  null.lo = log(1/1.1), null.hi = log(1.1),  pt.est = 2,  std.err = 0.8,  null.weights = 'uniform',  null.space = c(log(1/1.1), log(1.1)),  alt.weights = 'uniform',  alt.space = 2 + c(-1,1)*qnorm(1-0.05/2)*0.8,  interval.type = 'confidence',  interval.level = 0.05)

[1] 0.05949861

# false discovery risk with likelihood support level
> fdrisk(sgpvalue = 0,  null.lo = log(1/1.1), null.hi = log(1.1),  pt.est = 2,  std.err = 0.8,  null.weights = 'point',  null.space = 0,  alt.weights = 'uniform',  alt.space = 2 + c(-1,1)*qnorm(1-0.041/2)*0.8,  interval.type = 'likelihood',  interval.level = 1/8)

[1] 0.05055513

# false discovery risk with LSI and wider null hypothesis
> fdrisk(sgpvalue = 0,  null.lo = log(1/1.5), null.hi = log(1.5),  pt.est = 2.5,  std.err = 0.8,  null.weights = 'point',  null.space = 0,  alt.weights = 'uniform',  alt.space = 2.5 + c(-1,1)*qnorm(1-0.041/2)*0.8,  interval.type = 'likelihood',  interval.level = 1/8)

[1] 0.01688343

# can only calculate FDR or FCR for sgpvalue = 0 or 1:
> lb = 2 - qnorm(1-0.041/2)*0.8
> ub = 2 + qnorm(1-0.041/2)*0.8
> sgpv = sgpvalue(est.lo = lb, est.hi = ub, null.lo = log(1/1.5), null.hi = log(1.5))$p.delta
> sgpv

[1] 0.0248413

> fdrisk(sgpvalue = sgpv,  null.lo = log(1/1.5), null.hi = log(1.5),  pt.est = 2,  std.err = 0.8,  null.weights = 'point',  null.space = 0,  alt.weights = 'uniform',  alt.space = 2 + c(-1,1)*qnorm(1-0.041/2)*0.8,  interval.type = 'likelihood',  interval.level = 1/8)

     Error in fdrisk(sgpvalue = sgpv, null.lo = log(1/1.5), null.hi = log(1.5),  :
       sgpvalue must take a value of 0 or 1 to use fdrisk



}
\references{
Blume JD, D’Agostino McGowan L, Dupont WD, Greevy RA Jr (2018) Second-generation \emph{p}-values: Improved rigor, reproducibility, & transparency in statistical analyses. PLoS ONE 13(3): e0188299. https://doi.org/10.1371/journal.pone.0188299

Blume JD, Welty VF, Dupont WD, Greevy RA Jr. An Introduction to Second-Generation \emph{p}-values. The American Statistician. doi.
}
\seealso{
\code{\link{sgpvalue}}
}
