% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpv.R
\name{sgpv}
\alias{sgpv}
\title{Second Generation p-values}
\usage{
sgpv(data, estimate_lb = 1, estimate_ub = 2, null_lb = 3,
  null_ub = 4, epsilon = 1e-05)
}
\arguments{
\item{data}{A data.frame or data.table which includes interval bounds for the estimate and the null region.}

\item{estimate_lb}{Column number or variable name of the estimate interval lower bound.  Default = 1}

\item{estimate_ub}{Column number or variable name of the estimate interval upper bound.  Default = 2}

\item{null_lb}{Column number or variable name of the null interval lower bound.  Default = 3}

\item{null_ub}{Column number or variable name of the null interval upper bound.  Default = 4}

\item{epsilon}{A (usually) small scalar to denote positive but infinitesimally small p_δ.  Default = 1e-5.  Values that are infinitesimally close to 1 are assigned \code{1-epsilon}.}
}
\description{
This function calculates p_δ and δ-gap.
}
\examples{
sgpv()

}
\keyword{test_cases_sgpv}
