% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpv_plot_intervals.R
\name{sgpv_plot_intervals}
\alias{sgpv_plot_intervals}
\title{Plot intervals}
\usage{
sgpv_plot_intervals(data, estimate_lb = 1, estimate_ub = 2, null_lb = 3,
  null_ub = 4, yvar = NULL, shift = 0, est_col = "#d7191c80",
  null_col = "#2c7bb680")
}
\arguments{
\item{data}{A data.frame or data.table which includes interval bounds for the estimate and the null region.}

\item{estimate_lb}{Column number or variable name of the estimate interval lower bound.  Default = 1}

\item{estimate_ub}{Column number or variable name of the estimate interval upper bound.  Default = 2}

\item{null_lb}{Column number or variable name of the null interval lower bound.  Default = 3}

\item{null_ub}{Column number or variable name of the null interval upper bound.  Default = 4}

\item{yvar}{Column number or variable name of the column which indicates the y-value of the interval in the plot.  If NULL, intervals are ploted at integer values corresponding to the negative row number.}

\item{shift}{Scalar value indicating how much to separate null and estimate intervals vertically.  Value is relative to \code{yvar}.  Default is 0.}

\item{est_col}{Color of estimate intervals}

\item{null_col}{Color of null intervals}
}
\value{
Function invisibly returns a list with the bookend values used in the plot, the pch matrix used for the symbols, and the yvar matrix.
}
\description{
This function plots null and estimation intervals
}
\details{
Generates a plot with intervals.
}
\examples{
sgpv_test_cases() \%>\% sgpv_plot_intervals

}
\keyword{test_cases_sgpv}
