% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpvalue.R
\name{sgpvalue}
\alias{sgpvalue}
\title{Second-Generation p-values}
\usage{
sgpvalue(est.lo, est.hi, null.lo, null.hi, inf.correction = 1e-05,
  length.warn = TRUE)
}
\arguments{
\item{est.lo}{A (non-empty) numeric vector of lower bounds of interval estimates. These may be finite or they may be \code{-Inf} or \code{Inf}.}

\item{est.hi}{A (non-empty) numeric vector of upper bounds of interval estimates. These may be finite or they may be \code{-Inf} or \code{Inf}. Must be of same length as \code{est.lo}.}

\item{null.lo}{A (non-empty) numeric vector of lower bounds of null intervals. These may be finite or they may be \code{-Inf} or \code{Inf}.}

\item{null.hi}{A (non-empty) numeric vector of upper bounds of null intervals. These may be finite or they may be \code{-Inf} or \code{Inf}. Must be of same length as \code{null.lo}. If not of length 1, must be of same length as \code{est.lo} and \code{est.hi}.}

\item{inf.correction}{A (usually) small scalar to denote positive but infinitesimally small SGPV. Default = 1e-5. Values that are infinitesimally close to 1 are assigned \code{1-inf.correction}.}

\item{length.warn}{Warnings toggle. Default is TRUE.}
}
\value{
A list containing the following components:
\describe{
\item{\code{p.delta}}{the vector of calculated second-generation p-values}
\item{\code{delta.gap}}{the vector of calculated delta-gaps. Will be \code{NA} for those with corresponding second-generation p-values not equal to zero.}
}
}
\description{
This function calculates the Second-Generation p-value (SGPV) and associated delta gaps, as introduced in Blume et al. (2018).
}
\details{
[Details]
Note that the second-generation \emph{p}-value should be calculated on the "symmetric" scale. For example, if the parameter of interest is an odds ratio, the inputs \code{est.lo}, \code{est.hi}, \code{null.lo}, and \code{null.hi} should be on the scale of log odds ratios.
}
\examples{
## Single inputs
> sgpvalue(est.lo = log(1.05), est.hi = log(1.8), null.lo = log(1/1.1), null.hi = log(1.1))

$p.delta
[1] 0.1220227

$delta.gap
[1] NA

> sgpvalue(est.lo = log(1.3), est.hi = log(1.8), null.lo = log(1/1.1), null.hi = log(1.1))

$p.delta
[1] 0

$delta.gap
[1] 1.752741

> sgpvalue(est.lo = log(0.97), est.hi = log(1.02), null.lo = log(1/1.1), null.hi = log(1.1))

$p.delta
[1] 1

$delta.gap
[1] NA

## Vectors of inputs
> lb = c(log(1.05), log(1.3), log(0.97))
> ub = c(log(1.8), log(1.8), log(1.02))
> sgpvalue(est.lo = lb, est.hi = ub, null.lo = rep(log(1/1.1), 3), null.hi = rep(log(1.1), 3))

$p.delta
[1] 0.1220227 0.0000000 1.0000000

$delta.gap
[1] NA        1.752741  NA

## To get only the vector of second-generation p-values or only the vector of delta-gaps:
> sgpvalue(est.lo = lb, est.hi = ub, null.lo = rep(log(1/1.1), 3), null.hi = rep(log(1.1), 3))$p.delta

[1] 0.1220227 0.0000000 1.0000000

> sgpvalue(est.lo = lb, est.hi = ub, null.lo = rep(log(1/1.1), 3), null.hi = rep(log(1.1), 3))$delta.gap

[1] NA        1.752741  NA

## Interval bounds may be infinite:
> sgpvalue(est.lo = log(1.3), est.hi = Inf, null.lo = -Inf, null.hi = log(1.1))

$p.delta
[1] 0

$delta.gap
[1] 0.1670541

Warning message:
  In sgpvalue(est.lo = log(1.3), est.hi = Inf, null.lo = -Inf, null.hi = log(1.1)) :
    At least one interval has infinite length

> sgpvalue(est.lo = log(1.05), est.hi = Inf, null.lo = -Inf, null.hi = log(1.1))

$p.delta
[1] 1e-05

$delta.gap
[1] NA

Warning message:
    In sgpvalue(est.lo = log(1.05), est.hi = Inf, null.lo = -Inf, null.hi = log(1.1)) :
      At least one interval has infinite length


}
\references{
Blume JD, D’Agostino McGowan L, Dupont WD, Greevy RA Jr (2018) Second-generation \emph{p}-values: Improved rigor, reproducibility, & transparency in statistical analyses. PLoS ONE 13(3): e0188299. https://doi.org/10.1371/journal.pone.0188299

Blume JD, Welty VF, Dupont WD, Greevy RA Jr. An Introduction to Second-Generation \emph{p}-values. The American Statistician. doi.
}
\seealso{
\code{\link{fdrisk}}
}
